+++ consensus/src/lib.rs
@@ -1,6 +1,7 @@
 //! TBD
 
 pub mod fixed;
+pub mod mocks;
 
 use bytes::Bytes;
 use std::future::Future;
